refactor: Grouped related files (create-registries)
