+++ src/steps/create-registries/rename-component.ts
@@ -2,8 +2,10 @@ import {
   getBaseComponent,
   type TransformedEntityName,
 } from '../../utils/components.js';
-import { passComponentNameToBaseComponent } from './pass-component-name-to-base-component.js';
-import { updateReferences } from './update-references.js';
+import {
+  passComponentNameToBaseComponent,
+  updateReferences,
+} from './rename-component/index.js';
 
 type Data = {
   entity: TransformedEntityName;
