+++ src/steps/create-registries/rename-component/index.ts
@@ -1,0 +1,2 @@
+export * from './pass-component-name-to-base-component.js';
+export * from './update-references.js';
