+++ src/steps/create-registries/rename-component/update-references.ts
@@ -1,6 +1,6 @@
 import { AST } from '@codemod-utils/ast-javascript';
 
-import type { TransformedEntityName } from '../../utils/components.js';
+import type { TransformedEntityName } from '../../../utils/components.js';
 
 type Options = {
   baseComponentName: string;
